+++ src/realm/link_view.cpp
@@ -78,15 +78,15 @@ void LinkView::set(size_t link_ndx, size_t target_row_ndx)
     if (num_remaining > 0)
         return;
 
-    ColumnBase::CascadeState::row target_row;
+    CascadeState::row target_row;
     target_row.table_ndx = target_table.get_index_in_group();
     target_row.row_ndx   = old_target_row_ndx;
-    ColumnBase::CascadeState state;
+    CascadeState state;
     state.rows.push_back(target_row);
 
     typedef _impl::TableFriend tf;
     tf::cascade_break_backlinks_to(target_table, old_target_row_ndx, state); // Throws
-    tf::remove_backlink_broken_rows(target_table, state.rows); // Throws
+    tf::remove_backlink_broken_rows(target_table, state); // Throws
 }
 
 
@@ -148,15 +148,15 @@ void LinkView::remove(size_t link_ndx)
     if (num_remaining > 0)
         return;
 
-    ColumnBase::CascadeState::row target_row;
+    CascadeState::row target_row;
     target_row.table_ndx = target_table.get_index_in_group();
     target_row.row_ndx   = target_row_ndx;
-    ColumnBase::CascadeState state;
+    CascadeState state;
     state.rows.push_back(target_row);
 
     typedef _impl::TableFriend tf;
     tf::cascade_break_backlinks_to(target_table, target_row_ndx, state); // Throws
-    tf::remove_backlink_broken_rows(target_table, state.rows); // Throws
+    tf::remove_backlink_broken_rows(target_table, state); // Throws
 }
 
 
@@ -193,7 +193,7 @@ void LinkView::clear()
     }
 
     size_t origin_row_ndx = get_origin_row_index();
-    ColumnBase::CascadeState state;
+    CascadeState state;
     state.stop_on_link_list_column  = &m_origin_column;
     state.stop_on_link_list_row_ndx = origin_row_ndx;
 
@@ -206,11 +206,10 @@ void LinkView::clear()
         size_t num_remaining = target_table.get_num_strong_backlinks(target_row_ndx);
         if (num_remaining > 0)
             continue;
-        ColumnBase::CascadeState::row target_row;
+        CascadeState::row target_row;
         target_row.table_ndx = target_table.get_index_in_group();
         target_row.row_ndx   = target_row_ndx;
-        typedef ColumnBase::CascadeState::row_set::iterator iter;
-        iter i = std::upper_bound(state.rows.begin(), state.rows.end(), target_row);
+        auto i = std::upper_bound(state.rows.begin(), state.rows.end(), target_row);
         // This target row cannot already be in state.rows
         REALM_ASSERT(i == state.rows.begin() || i[-1] != target_row);
         state.rows.insert(i, target_row);
@@ -220,7 +219,7 @@ void LinkView::clear()
     bool broken_reciprocal_backlinks = true;
     do_clear(broken_reciprocal_backlinks); // Throws
 
-    tf::remove_backlink_broken_rows(*m_origin_table, state.rows); // Throws
+    tf::remove_backlink_broken_rows(*m_origin_table, state); // Throws
 }
 
 
@@ -308,18 +307,9 @@ void LinkView::remove_all_target_rows()
 {
     REALM_ASSERT(is_attached());
 
-    Table& target_table = get_target_table();
-
-    // Delete all rows targeted by links. We have to keep checking the
-    // size as the list may contain multiple links to the same row, so
-    // one delete could remove multiple entries.
-    while (size_t count = size()) {
-        size_t last_link_ndx = count-1;
-        size_t target_row_ndx = m_row_indexes.get(last_link_ndx);
-
-        // Deleting the target row will automatically remove all links
-        // to it. So we do not have to manually remove the deleted link
-        target_table.move_last_over(target_row_ndx);
+    if (m_row_indexes.is_attached()) {
+        Table& target_table = get_target_table();
+        target_table.move_last_over(m_row_indexes);
     }
 }